+++ public/modules/automation/jira-protocol-wizard.js
@@ -65,7 +65,8 @@
             label: "Tylko aktywni użytkownicy",
             required: false,
             control: "switch",
-            defaultValue: "true"
+            defaultValue: "true",
+            inlineControl: true
         };
         var search = {
             name: "JiraUserSearch",
@@ -77,6 +78,7 @@
         jiraUser = copy(jiraUser);
         jiraUser.label = "Użytkownicy";
         jiraUser.description = "";
+        jiraUser.hideLabel = true;
         jiraUser.listMode = true;
         jiraUser.submitOnDoubleClick = true;
         jiraUser.searchVariable = "JiraUserSearch";
@@ -84,6 +86,7 @@
         asset = copy(asset);
         asset.label = "Sprzęt";
         asset.description = "";
+        asset.hideLabel = true;
         asset.control = "assetmulti";
         transfer = copy(transfer);
         transfer.description = "";
@@ -100,7 +103,7 @@
         delete itPerson.defaultValue;
 
         var userStep = stepItem(item, "Jira Asset Protocol - User", "", [activeOnly, search, jiraUser]);
-        var assetStep = stepItem(item, "Sprzęt", "", [asset]);
+        var assetStep = stepItem(item, "Sprzęt do protokołu", "", [asset]);
         var protocolStep = stepItem(item, "Jira Asset Protocol - Protocol", "", [transfer, itPerson]);
 
         return runStep(options, userStep, {}, "Next").then(function (userValues) {